Add special rules for MinGW
Some (all?) variants of MinGW do not have a working 'fmt'.
This is not indispensable to the rule working, it makes it just look good.

The sed -e 's/\\\\...' rule appears necessary when executing the
command within 'make'. This is definitely a bug. If I execute the
command within the Windows CMD, I need the extra \\. If I execute the
command within bash or sh, I do not.